eventmanager class
